refactor(paywall): register the card as platform-paywall
The card reads the platform billing snapshot, platform roles and the proxy's
subscription_required flag, so it is a platform component. Under the generic key
'paywall' another provider (e.g. OpenRouter) emitting the same key would have its
composer gated by the wrong org's billing. Rename key, constant, component, hook and
files to platform-paywall-*; a non-platform provider registers its own row.
