+++ src/renderer/components/provider-error/platform-paywall-cta.ts
