+++ src/renderer/components/provider-error/provider-error-registry.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, it } from 'vitest'
 
 import { ProviderErrorCard } from '@renderer/components/ui/provider-error-card'
 
-import { PaywallCard } from './paywall-card'
+import { PlatformPaywallCard } from './platform-paywall-card'
 import { resolveProviderError } from './provider-error-registry'
 
 const base = { severity: 'error' as const, message: 'x', icon: 'info' }
@@ -22,8 +22,8 @@ describe('resolveProviderError', () => {
     expect(resolveProviderError({ ...base, component: 'not-registered' }).Component).toBe(ProviderErrorCard)
   })
 
-  it('resolves the paywall key to PaywallCard', () => {
-    expect(resolveProviderError({ ...base, component: 'paywall' }).Component).toBe(PaywallCard)
+  it('resolves the platform-paywall key to PlatformPaywallCard', () => {
+    expect(resolveProviderError({ ...base, component: 'platform-paywall' }).Component).toBe(PlatformPaywallCard)
   })
 
   it('passes placement through', () => {
